+++ Examples/Extensions/Texture2D/BlendTextureExample.cs
@@ -1,9 +1,9 @@
 ﻿using UnityEngine;
 using UnityEngine.UI;
 
-namespace UniSharper.Examples.Extensions.Camera
+namespace UniSharper.Examples.Extensions.Texture2D
 {
-    internal class CaptureScreenshotTextureExample : MonoBehaviour
+    internal class BlendTextureExample : MonoBehaviour
     {
         [SerializeField]
         private UnityEngine.Camera mainCamera = null;
@@ -24,11 +24,20 @@ namespace UniSharper.Examples.Extensions.Camera
         {
             rawImage1.texture = mainCamera.CaptureScreenshotTexture();
             rawImage2.texture = mainCamera.CaptureScreenshotTexture(new Rect(Screen.width / 4f, Screen.height / 4f, Screen.width / 2f, Screen.height / 2f));
-            rawImage3.texture = mainCamera.CaptureScreenshotTexture(new Rect(0, 0, Screen.width / 2f, Screen.height / 2f));
             rawImage4.texture = mainCamera.CaptureScreenshotTexture(new Rect(0, Screen.height / 2f, Screen.width, Screen.height / 2f));
             rawImage4.rectTransform.sizeDelta = new Vector2(rawImage4.rectTransform.sizeDelta.x, rawImage4.rectTransform.sizeDelta.x / (rawImage4.mainTexture.width / rawImage4.mainTexture.height));
+            BlendTexture();
+        }
+
+        private void BlendTexture()
+        {
+            var texture1 = rawImage1.texture as UnityEngine.Texture2D;
+            var texture2 = rawImage4.texture as UnityEngine.Texture2D;
+
+            if (!texture1 || !texture2)
+                return;
+
+            rawImage3.texture = texture1.BlendTexture(texture2, new Vector2Int(100, 150));
         }
     }
 }
-
-